[SPARC64]: Make access_ok() more palatable to gcc-4.x
Also, kill __{user,kernel}_ok() as they are unused
macros on sparc64.

+++ arch/sparc64/kernel/signal32.c
@@ -1102,7 +1102,7 @@ asmlinkage int svr4_setcontext(svr4_ucontext_t __user *c, struct pt_regs *regs)
 		goto sigsegv;
 	}
 
-	if (!__access_ok((unsigned long)c, sizeof(*c))) {
+	if (!__access_ok(c, sizeof(*c))) {
 		/* Miguel, add nice debugging msg _here_. ;-) */
 		goto sigsegv;
 	}
